Fix class names
